First stab at making drift detection pluggable. This is not complete by any means.
